pyproject.toml
[project]
name = "iap-2026-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "openai>=1.0.0",
    "notion-client>=2.0.0",
    "replicate>=0.10.0",
    "python-telegram-bot>=21.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "sqlite-vec>=0.1.1",
    "fastembed>=0.4.0",
    "numpy>=1.24.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.13",
]
